+++ hi-ml/src/health_ml/__init__.py
@@ -2,7 +2,7 @@
 #  Copyright (c) Microsoft Corporation. All rights reserved.
 #  Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
 #  ------------------------------------------------------------------------------------------
-from health_ml.ml_runner import MLRunner
+from health_ml.training_runner import MLRunner
 from health_ml.runner import Runner
